[feat] 교수페이지 데이터 분리, 학과페이지 데이터 분리
