+++ src/pages/Departments/Department/Course.jsx
@@ -1,0 +1,22 @@
+import "./Department.css"
+const Course = () => {
+    return (
+        <div className="Course">
+            <span>진로</span>
+        <pre className="Another">{`
+국내외 수많은 정보통신, 컴퓨터, 인터넷, 멀티미디어, 게임, 애니메이션, 전자상거래 관련기업, 산업체, 금융기관, 국공립 연구소에 진출할 수 있다.
+또한 졸업 후에는 대학원에 진학하거나 유학을 가고 벤처기업을 창업하는 졸업생이 점차 늘어나고 있다.
+        `}</pre>
+        <pre className="Another">{`
+향후 정보화 사회가 본격적으로 도래함에 따라 컴퓨터공학 전공자의 수요가 폭발적으로 증가되고 있어, 창의력과 전문성을 가진 컴퓨터공학
+전공자들이 국가경쟁력 제고에 큰 역할을 할 것으로 기대된다.
+        `}</pre>
+        <pre className="Anotherlast">{`
+국가적, 사회적, 시대적 요구에 비추어 보아 컴퓨터공학 전공에서 교육하고 다루는 기술은 산업발전의 핵심이며 주체이며,
+이러한 점에서 컴퓨터공학 전공의 향후 발전 전망은 매우 밝고 무한하다고 할 수 있다.
+        `}</pre>
+        </div>
+    )
+  };
+  
+  export default Course
