+++ src/pages/Departments/Department/Department.jsx
@@ -9,41 +9,6 @@ const Department = () => {
 기본적인 컴퓨터관련 기술에 대한 체계적인 학습과 급변하는 정보기술 발전에 부합하는 스마트모바일 컴퓨팅, 정보보호 및 보안,
 인터넷 응용기술, 인공지능, 병렬컴퓨터, 차세대 데이터베이스, 자연어 처리, 컴퓨터구조설계, 임베디드 시스템 등에 대한 학문을 연구한다.
       `}</pre>
-      <span>인재상</span>
-      <p>사랑과 협동의 정신을 바탕으로, 미래 정보화와 첨단 산업사회를 창의적으로 선도하여 지역과 인류 사회의 발전에 기여하는, 전문 인력을 양성한다.</p>
-      <ul>
-        <li>전문 기술인으로서의 사명 의식과 공동체 속에서 더불어 살아가는 지혜를 갖춘 아름다운 리더를 양성한다.</li>
-        <li>컴퓨터공학의 전문 이론을 탐구하여, 정보화 사회를 선도하는 지식 창출 능력을 갖춘 전문인을 양성한다.</li>
-        <li>첨단산업기술 사회에 필요한 실용적 기술을 연마하여, 산업 현장에 적용할 수 잇는 창의적 응용력을 갖춘 기술인을 양성한다.</li>
-        <li>컴퓨터공학의 전문 기술을 바탕으로 지역과 인류 사회의 발전에 기여하는 지성인을 양성한다.</li>
-      </ul>
-      <span>특성화 방향</span>
-      <pre>{`
-우리 학과는 경인지역의 다양한 기술 집약형 IT기업 클러스터 내에 위치한 지리적 강점을 활용하여 기업들의 요구를 반영하는 맞춤형 산학연계 교육체계를
-구축함으로써, 실무 중심형 IT기술 특히 인공지능, 사물인터넷 및 빅데이터 처리 소프트웨어 설계능력에 특화된 고급 전문 인력을 양성한다.
-이를 위해 교육과정은 다양한 학문 분야와 IT기술의 융합 기회를 제공하고, 연구실 중심의 개인별 밀착형 교육 체계를 시행한다.
-      `}</pre>
-      <span>진로</span>
-        <pre className="Another">{`
-국내외 수많은 정보통신, 컴퓨터, 인터넷, 멀티미디어, 게임, 애니메이션, 전자상거래 관련기업, 산업체, 금융기관, 국공립 연구소에 진출할 수 있다.
-또한 졸업 후에는 대학원에 진학하거나 유학을 가고 벤처기업을 창업하는 졸업생이 점차 늘어나고 있다.
-        `}</pre>
-        <pre className="Another">{`
-향후 정보화 사회가 본격적으로 도래함에 따라 컴퓨터공학 전공자의 수요가 폭발적으로 증가되고 있어, 창의력과 전문성을 가진 컴퓨터공학
-전공자들이 국가경쟁력 제고에 큰 역할을 할 것으로 기대된다.
-        `}</pre>
-        <pre className="Anotherlast">{`
-국가적, 사회적, 시대적 요구에 비추어 보아 컴퓨터공학 전공에서 교육하고 다루는 기술은 산업발전의 핵심이며 주체이며,
-이러한 점에서 컴퓨터공학 전공의 향후 발전 전망은 매우 밝고 무한하다고 할 수 있다.
-        `}</pre>
-        <span>관련 자격증</span>
-        <ul>
-          <li>컴퓨터시스템응용기사</li>
-          <li>전자계산기조직응용기사</li>
-          <li>정보처리기사</li>
-          <li>정보보안기산</li>
-          <li>게임프로그래밍전문가</li>
-        </ul>
     </div>
   )
 };
