+++ src/pages/Departments/Department/Direction.jsx
@@ -1,0 +1,15 @@
+import "./Department.css"
+const Direction = () => {
+    return (
+        <div className="Direction">
+            <span>특성화 방향</span>
+      <pre>{`
+우리 학과는 경인지역의 다양한 기술 집약형 IT기업 클러스터 내에 위치한 지리적 강점을 활용하여 기업들의 요구를 반영하는 맞춤형 산학연계 교육체계를
+구축함으로써, 실무 중심형 IT기술 특히 인공지능, 사물인터넷 및 빅데이터 처리 소프트웨어 설계능력에 특화된 고급 전문 인력을 양성한다.
+이를 위해 교육과정은 다양한 학문 분야와 IT기술의 융합 기회를 제공하고, 연구실 중심의 개인별 밀착형 교육 체계를 시행한다.
+      `}</pre>
+        </div>
+    )
+  };
+  
+  export default Direction
