+++ src/pages/Departments/Department/IdolTalent.jsx
@@ -1,0 +1,17 @@
+import "./Department.css"
+const IdolTalent = () => {
+    return (
+        <div className="IdolTalent">
+            <span>인재상</span>
+      <p>사랑과 협동의 정신을 바탕으로, 미래 정보화와 첨단 산업사회를 창의적으로 선도하여 지역과 인류 사회의 발전에 기여하는, 전문 인력을 양성한다.</p>
+      <ul>
+        <li>전문 기술인으로서의 사명 의식과 공동체 속에서 더불어 살아가는 지혜를 갖춘 아름다운 리더를 양성한다.</li>
+        <li>컴퓨터공학의 전문 이론을 탐구하여, 정보화 사회를 선도하는 지식 창출 능력을 갖춘 전문인을 양성한다.</li>
+        <li>첨단산업기술 사회에 필요한 실용적 기술을 연마하여, 산업 현장에 적용할 수 잇는 창의적 응용력을 갖춘 기술인을 양성한다.</li>
+        <li>컴퓨터공학의 전문 기술을 바탕으로 지역과 인류 사회의 발전에 기여하는 지성인을 양성한다.</li>
+      </ul>
+        </div>
+    )
+  };
+  
+  export default IdolTalent
