+++ src/pages/Departments/Department/Qualifications.jsx
@@ -1,0 +1,17 @@
+import "./Department.css"
+const Qualifications = () => {
+    return (
+        <div className="Qualifications">
+            <span>관련 자격증</span>
+        <ul>
+          <li>컴퓨터시스템응용기사</li>
+          <li>전자계산기조직응용기사</li>
+          <li>정보처리기사</li>
+          <li>정보보안기산</li>
+          <li>게임프로그래밍전문가</li>
+        </ul>
+        </div>
+    )
+  };
+  
+  export default Qualifications
