+++ src/pages/Departments/DepartmentInfo/DepartmentInfo.jsx
@@ -2,6 +2,10 @@
 import Footer from "../../../components/Footer/Footer";
 import Nav from "../../../components/Nav/Nav";
 import Department from "../Department/Department";
+import IdolTalent from "../Department/IdolTalent";
+import Direction from "../Department/Direction";
+import Course from "../Department/Course";
+import Qualifications from "../Department/Qualifications";
 import "./DepartmentInfo.css"
 
 const DepartmentInfo = () => {
@@ -14,6 +18,10 @@ const DepartmentInfo = () => {
       <p style={{ display: "inline-block", fontSize: "20px", fontWeight: "bold" }}>학과 안내</p>
       </div>
       <Department/>
+      <IdolTalent/>
+      <Direction/>
+      <Course/>
+      <Qualifications/>
       <Footer/>
     </div>
   )
