+++ src/pages/Professors/Professors/Professorsdata.js
@@ -1,0 +1,48 @@
+const professors = [
+    {
+        사진: "../public/image 32.png",
+        이름: "이상홍",
+        직책: "교수(학과장)",
+        연구실: "아리관 308호",
+        연락처: "031-467-0876",
+        이메일: "shleedosa@anyang.ac.kr",
+        학위: "대학원 박사졸업 - 경원대학교 전자계산학"
+    },
+    {
+        사진: "../public/image 27.png",
+        이름: "권희용",
+        직책: "교수",
+        연구실: "아리관 310호",
+        연락처: "031-467-0878",
+        이메일: "hykwon@anyang.ac.kr",
+        학위: "대학원 박사졸업 - 서울대학교 컴퓨터공학"
+    },
+    {
+        사진: "../public/image 31.png",
+        이름: "김성규",
+        직책: "교수",
+        연구실: "아리관 312호",
+        연락처: "031-467-0879",
+        이메일: "sgkim@anyang.ac.kr",
+        학위: "대학원 박사졸업 - 노스웨스턴대학교 전산학"
+    },
+    {
+        사진: "../public/image 30.png",
+        이름: "박성순",
+        직책: "교수",
+        연구실: "아리관 309호",
+        연락처: "031-467-0877",
+        이메일: "sspark@anyang.ac.kr",
+        학위: "대학원 박사졸업 - 고려대학교 컴퓨터과학"
+    },
+    {
+        사진: "../public/image 26.png",
+        이름: "하은용",
+        직책: "교수",
+        연구실: "아리관 311호",
+        연락처: "031-467-0873",
+        이메일: "eyha@anyang.ac.kr",
+        학위: "대학원 박사졸업 - 서울대학교 컴퓨터공학"
+    },
+];
+export default professors
